[lib] Use ‘size_t’ for memory allocation state.
* src/libserveez/alloc.c (allocated_bytes)
(allocated_blocks): Convert type to ‘size_t’.
(svz_get_curalloc): Likewise for arg.
* src/libserveez/alloc.h (svz_get_curalloc): Update func decl.

* src/ctrl-server/control-proto.c (svz_sock_printf):
Update type of arg to ‘svz_get_curalloc’.
* src/serveez.c (guile_entry): Likewise.

* test/btdt.c (array_main, hash_main, codec_main):
Update type of arg to ‘svz_get_curalloc’.

+++ src/ctrl-server/control-proto.c
@@ -576,7 +576,7 @@ ctrl_stat (svz_socket_t *sock, int flag, char *arg)
   svz_sock_printf (sock, " * uptime is %s\r\n", uptime (ut));
 #if ENABLE_DEBUG
   {
-    unsigned int cur[2];
+    size_t cur[2];
 
     svz_get_curalloc (cur);
     svz_sock_printf (sock, " * %d bytes of memory in %d blocks allocated\r\n",
